Fix activeElement
https://github.com/lightpanda-io/browser/pull/742 was only correct in isolation.

The problem is that both html/document and dom/document are now attempting to
attach state to the same *parser.Node. This is a problem because they're two
different types of state. If an HTMLDocument is attached to the node, and you
then try to access that as a DOMDocument, it'll crash.

I moved the dom/document state (active_element) into the html/document.

The other option is to introduce something like src/browser/state/document.zig
and have both HTMLDocument and DOMDocument share/load that.